Starting with this version only use the TDM-GCC-64 toolchain, it uses mingw32 and can compile for both, 32-bit and 64-bit.
On some Windows versions negative time value in 'date-list' for dates
before 1970-1-1 are not handled by gmtime() and could crash newLISP.
Now an error 44 is thrown when gmtime() cannot handle negative values.

Fixed 'det' and 'mat' for Ubuntu Linux when using wrong dimensions.
Tested on OSX, OpenBSD, FreeBSD Ubuntu Linux, Windows.

Changes in several files for compiling a 64-bit version for Windows
using mingw64. Thanks to Shigeru Kobayashi for doing this work.

Filenames win32xxx.c change to win-xxx.c.

The 'dolist' function now also works on arrays.

The 'net-eval' functions now return the result of the last expression
in the string argument, not the first.
    (net-eval host port "(+ 3 4) (* 3 4)" => 12 ; previously 7
In previous versions, although all expressions in 'net-eval' would be
evaluated, the result of the first was returned. The possibility
of multiple expressions in a string was not documented. The new
behaviour is more in line with the behaviour of 'eval-string' and
'load'.

Fixed a bug on Windows when APACHE was starting newlisp and no
NEWLISPDIR wnd PROGRAMFILES environment variables were defined.

New makefiles for Windows:
    makefile_mingw64 makefile_mingw64_ffi
    makefile_mingw64_utf8 makefile_mingw64_utf8_ffi
    makefile_mingw64dll makefile_mingw64dll_ffi
    makefile_mingw64dll_utf8 makefile_mingw64dll_utf8_ffi
are working. Tested using tdm gcc 64 (MinGW based).

The configure utility now uses the mingw64 makefiles if LLP64 is present.
The configure-alt utility still only compiles for 32-bit.

Both, 32-bit and 64-bit newLISP on Windows, now have ostype "Windows".
In modules canvas.lsp and gsl.lsp ostype has been changed to "Windows"
from "Win32". Same change in many qa-* files and web applications.

C type cleanup for string size_t length in some C-functions.

Server mode -http did not reject net-eval requests.

'net-eval' connection timeout changed to 15 secs from 60 secs.
Processing timeout still set by user in 'net-eval' parameter.

Added HTTP_AUTHORIZATION to environment variables in HTTP server.

Additional syntax for 'date-value' -> (date-value list-date-time).
Additional syntax (date-list) now same as (date-list (date-value)).

The 'trim' function without trim-char parameters will trim all
white-space characters, not only spaces.

Fixed 'expand' bug when evaluating binding lists.

'dump' now returns the number of cells dumped not 'true'

New 'trace' syntax: (trace device-no) prints an evaluation trace to
device-no. Example:
    (trace (open "trace.txt" "w")) ; open trace file (or "a")
    ... ; one or more expressions to evaluate
    (trace nil) ; close trace file

Stack trace in error message changed from function name to full
expression.

tmp directory is now defined on newLISP startup. For UNIX it is /tmp
for Windows it is taken from TMP environment variable or assumed /tmp.
This directory is used by the 'share' functions and the built-in http
server.

When char* in libffi returns 0 (NULL) then an error message will be
thrown: "cannot convert NULL to string".

A new newLISP library call (newlispLibConsole 1) forces console output
for stdout instead of writing it to the return string of newlispEvalStr.
The same also enables console input on stdin.

The 'json-parse' function now handles 64-bit number also in 32-bit versions
of newLISP.

Fixed shape sign error in canvas.lsp for bezier curves in shapes.

+++ nl-liststr.c
@@ -19,7 +19,6 @@
 */
 
 #include "newlisp.h"
-#include "pcre.h"
 #include "protos.h"
 
 extern CELL * lastCellCopied;
@@ -1062,7 +1061,7 @@ if(keyCell->type == CELL_STRING && next->type == CELL_STRING)
     if(options == -1)
         found = searchBuffer(str + offset, size - offset, key, keyCell->aux - 1, TRUE);
     else
-        found = searchBufferRegex(str, (int)offset, key, (int)size, options, NULL) - offset;
+        found = searchBufferRegex(str, offset, key, size, options, NULL) - offset;
     if(found < 0) return(nilCell);
     }
 else
@@ -1135,8 +1134,8 @@ CELL * cell = NULL;
 UINT * resultIdxSave;
 jmp_buf errorJumpSave;
 int errNo; 
-int len;
-int offset = 0;
+size_t len;
+size_t offset = 0;
 
 exprCell = params;
 if((params = params->next) != nilCell)
@@ -1157,7 +1156,7 @@ if(exprCell != nilCell)
         }   
     }
 
-while( (findPos = searchBufferRegex(str, offset, pattern, (int)size, options, &len)) 
+while( (findPos = searchBufferRegex(str, offset, pattern, size, options, &len)) 
        != -1)
     {
     countCell->contents++;
@@ -1406,7 +1405,7 @@ char * key;
 char * keydollar;
 INT options = -1;
 size_t slen, pos;
-int klen;
+size_t klen;
 CELL * cell, * list;
 
 cell = params->next;
